Переносит подсчет статуса
